Revert "Merge branch 'Mahmoud' into Capa"
This reverts commit 03308374554f1d963a407c33ceeef1660c9ae0a5, reversing
changes made to 0720577ee6c1940d4955e93c42d4fc525704c972.

+++ chickenEggs/interfaces/Game/singleGameObject.java
@@ -2,8 +2,8 @@ package chickenEggs.interfaces.Game;
 import chickenEggs.interfaces.drawable;
 
 public class singleGameObject extends drawable{
-    public float xminGame,xmaxGame,yminGame,ymaxGame,speed=5;
-    public singleGameObject(int x, int y, int w, int h, int path, float xminGame, float xmaxGame, float yminGame, float ymaxGame){
+    public int xminGame,xmaxGame,yminGame,ymaxGame,speed=5;
+    public singleGameObject(int x, int y, int w, int h, int path, int xminGame, int xmaxGame, int yminGame, int ymaxGame){
         super(x,y,w,h,path);
         this.xminGame=xminGame;
         this.xmaxGame=xmaxGame;
